update: move histogram to another src
