+++ dpnp/dpnp_iface.py
@@ -64,7 +64,6 @@ __all__ = [
 from dpnp.dpnp_iface_arraycreation import *
 from dpnp.dpnp_iface_bitwise import *
 from dpnp.dpnp_iface_counting import *
-from dpnp.dpnp_iface_histograms import *
 from dpnp.dpnp_iface_indexing import *
 from dpnp.dpnp_iface_libmath import *
 from dpnp.dpnp_iface_linearalgebra import *
@@ -79,7 +78,6 @@ from dpnp.dpnp_iface_trigonometric import *
 from dpnp.dpnp_iface_arraycreation import __all__ as __all__arraycreation
 from dpnp.dpnp_iface_bitwise import __all__ as __all__bitwise
 from dpnp.dpnp_iface_counting import __all__ as __all__counting
-from dpnp.dpnp_iface_histograms import __all__ as __all__histograms
 from dpnp.dpnp_iface_indexing import __all__ as __all__indexing
 from dpnp.dpnp_iface_libmath import __all__ as __all__libmath
 from dpnp.dpnp_iface_linearalgebra import __all__ as __all__linearalgebra
@@ -94,7 +92,6 @@ from dpnp.dpnp_iface_trigonometric import __all__ as __all__trigonometric
 __all__ += __all__arraycreation
 __all__ += __all__bitwise
 __all__ += __all__counting
-__all__ += __all__histograms
 __all__ += __all__indexing
 __all__ += __all__libmath
 __all__ += __all__linearalgebra
