--- dpnp/dpnp_iface_histograms.py
@@ -1,82 +1,0 @@
-# cython: language_level=3
-# distutils: language = c++
-# -*- coding: utf-8 -*-
-# *****************************************************************************
-# Copyright (c) 2016-2020, Intel Corporation
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-# - Redistributions of source code must retain the above copyright notice,
-#   this list of conditions and the following disclaimer.
-# - Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
-# THE POSSIBILITY OF SUCH DAMAGE.
-# *****************************************************************************
-
-"""
-Interface of the histogram-related functions of the dpnp
-
-Notes
------
-This module is a face or public interface file for the library
-it contains:
- - Interface functions
- - documentation for the functions
- - The functions parameters check
-
-"""
-
-
-from dpnp.dpnp_algo import *
-from dpnp.dpnp_utils import *
-
-import dpnp
-import numpy
-
-
-__all__ = [
-    'histogram',
-]
-
-
-def histogram(a, bins=10, range=None, normed=None, weights=None, density=None):
-    """
-    Compute the histogram of a dataset.
-
-    For full documentation refer to :obj:`numpy.histogram`.
-
-    Examples
-    --------
-    >>> import dpnp
-    >>> dpnp.histogram([1, 2, 1], bins=[0, 1, 2, 3])
-    (array([0, 2, 1]), array([0, 1, 2, 3]))
-    >>> dpnp.histogram(dpnp.arange(4), bins=dpnp.arange(5), density=True)
-    (array([0.25, 0.25, 0.25, 0.25]), array([0, 1, 2, 3, 4]))
-    >>> dpnp.histogram([[1, 2, 1], [1, 0, 1]], bins=[0,1,2,3])
-    (array([1, 4, 1]), array([0, 1, 2, 3]))
-    >>> a = dpnp.arange(5)
-    >>> hist, bin_edges = dpnp.histogram(a, density=True)
-    >>> hist
-    array([0.5, 0. , 0.5, 0. , 0. , 0.5, 0. , 0.5, 0. , 0.5])
-    >>> hist.sum()
-    2.4999999999999996
-    >>> res = dpnp.sum(hist * dpnp.diff(bin_edges))
-    >>> print(res)
-    1.0
-
-    """
-
-    return call_origin(numpy.histogram, a=a, bins=bins, range=range, normed=normed, weights=weights, density=density)
