+++ dpnp/dpnp_iface_statistics.py
@@ -54,6 +54,7 @@ __all__ = [
     'bincount',
     'correlate',
     'cov',
+    'histogram',
     'max',
     'mean',
     'median',
@@ -293,6 +294,33 @@ def cov(x1, y=None, rowvar=True, bias=False, ddof=None, fweights=None, aweights=
             return dpnp_cov(x1)
 
     return call_origin(numpy.cov, x1, y, rowvar, bias, ddof, fweights, aweights)
+
+
+def histogram(a, bins=10, range=None, normed=None, weights=None, density=None):
+    """
+    Compute the histogram of a dataset.
+    For full documentation refer to :obj:`numpy.histogram`.
+    Examples
+    --------
+    >>> import dpnp
+    >>> dpnp.histogram([1, 2, 1], bins=[0, 1, 2, 3])
+    (array([0, 2, 1]), array([0, 1, 2, 3]))
+    >>> dpnp.histogram(dpnp.arange(4), bins=dpnp.arange(5), density=True)
+    (array([0.25, 0.25, 0.25, 0.25]), array([0, 1, 2, 3, 4]))
+    >>> dpnp.histogram([[1, 2, 1], [1, 0, 1]], bins=[0,1,2,3])
+    (array([1, 4, 1]), array([0, 1, 2, 3]))
+    >>> a = dpnp.arange(5)
+    >>> hist, bin_edges = dpnp.histogram(a, density=True)
+    >>> hist
+    array([0.5, 0. , 0.5, 0. , 0. , 0.5, 0. , 0.5, 0. , 0.5])
+    >>> hist.sum()
+    2.4999999999999996
+    >>> res = dpnp.sum(hist * dpnp.diff(bin_edges))
+    >>> print(res)
+    1.0
+    """
+
+    return call_origin(numpy.histogram, a=a, bins=bins, range=range, normed=normed, weights=weights, density=density)
 
 
 def max(x1, axis=None, out=None, keepdims=False, initial=None, where=True):
